chỉnh sửa chút

+++ src/main/java/com/codegym/ecommercemanage/repository/CategoryRepository.java
@@ -5,5 +5,5 @@ import com.codegym.ecommercemanage.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 }
